use libgpiod
This is a major change as we don't use the sysfs interface anymore. This
was deprecated in 4.8. This also allows us to autodetect the GPIOs by
its names. Now three different configuration methods are supported.
Either the pins are automatically detected by their common names, the
user might supply its own names or the pins might be specified by the
gpiochip and offset.

Signed-off-by: Michael Walle <michael.walle@kontron.com>